src/requirements.txt
--trusted-host 44.211.201.219:8080
--extra-index-url http://44.211.201.219:8080

aiokafka==0.7.2
anyio==3.6.1;
asyncpg==0.26.0; 
certifi==2022.6.15;
cffi==1.15.1; 
click==8.1.3; 
colorama==0.4.5; 
cryptography==37.0.4; 
dependency-injector==4.40.0
ecdsa==0.18.0;
enowshop-models==0.4.33
fastapi==0.79.0; 
greenlet==1.1.2; 
h11==0.12.0; 
httpcore==0.15.0; 
httptools==0.4.0; 
httpx==0.23.0; 
idna==3.3
importlib-metadata==4.12.0; 
jwt==1.3.1; 
kafka-python==2.0.2
psycopg2-binary==2.9.3; 
pyasn1==0.4.8; 
pycparser==2.21; 
pycpfcnpj==1.7.0
pydantic==1.9.1; 
python-dotenv==0.20.0; 
python-jose==3.3.0
pyyaml==6.0; 
rfc3986==1.5.0; 
rsa==4.9; 
six==1.16.0; 
sniffio==1.2.0; 
sqlalchemy==1.4.39; 
starlette==0.19.1;
typing-extensions==4.3.0; 
uvicorn==0.18.3; 
uvloop==0.16.0; 
watchfiles==0.16.1;
websockets==10.3; 
zipp==3.8.1; 
